refactor(app): split startup and lifecycle mixins
Move notification activation, startup guidance, keyboard shortcuts, and update/tray/background lifecycle helpers into focused mixins while preserving AccessiWeatherApp as the main entry point.

Tested: app regression suite 141 passed; uv run ruff check src tests scripts; uv run pyright

Confidence: high
